Updating dependencies
Updated the Gemfile.lock to take care of the security notification
from github regarding nokogiri. Updated the vagrant file to only
create a VM with 512mb of ram as 2gb was excessive for a static
site. Tried to go lower but found that the VM was blowing up while
trying to set it up. Updated the cookbook for the VM so that it does
not install foreman by default when installing ruby as it is not
needed to run the site.